README.adoc: migrate from freenode to Libera
Also move logs from the shutting-down logbot.info (per
https://freenode.logbot.info/_shutdown) to irclog.whitequark.org.

See https://lwn.net/Articles/856543/ for pointers to more information on
the reasons behind this move.

I feel ashamed that this is the first commit to the repository in over
a year, and humbly resolve to feel even more ashamed if I don’t manage
any more before the year is out.